requirements.txt
annotated-types==0.7.0
dependency-injector==4.48.2
greenlet==3.2.4
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg==3.2.11
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
